Move common code within builders to a builders module
